+++ libamu/xutil.c
@@ -297,7 +297,7 @@ expand_error(const char *f, char *e, size_t maxlen)
   int error = errno;
   int len = 0;
 
-  for (p = f, q = e; (*q = *p) && len < maxlen; len++, q++, p++) {
+  for (p = f, q = e; (*q = *p) && (size_t) len < maxlen; len++, q++, p++) {
     if (p[0] == '%' && p[1] == 'm') {
       xstrlcpy(q, strerror(error), maxlen);
       len += strlen(q) - 1;
